requirements.txt
Django==1.5.4
amqp==1.0.13
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.34
blinker==1.3
celery==3.0.24
distribute==0.6.34
django-celery==3.0.23
django-celery-with-mongodb==3.0
flake8==2.0
ipdb==0.8
ipython==1.1.0
kombu==2.5.15
mccabe==0.2.1
mongoengine==0.8.4
oauthlib==0.6.0
pep8==1.4.6
pyflakes==0.7.3
pyjade==2.0.2
pylibmc==1.2.3
pymongo==2.6.3
python-dateutil==2.1
python-openid==2.2.5
python-social-auth==0.1.14
pytz==2013.7
requests==2.0.0
requests-oauthlib==0.4.0
six==1.4.1
wsgiref==0.1.2
